+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added One Minus float extension to do one minus (1-x) operations.
+
 ## [0.1.4] - 2022-05-03
 
 ### Fixed